requirements.txt
Django==1.7
Pillow==2.6.1
django-ckeditor-updated==4.4.4
gunicorn==19.1.1
